requirements.txt
DateTime==4.4
Flask==2.0.3
gunicorn==20.1.0
Jinja2==3.1.2
sqla-wrapper==4.200628
uuid==1.30
psycopg2-binary==2.9.3
redis==4.3.4
sendgrid==6.9.7
python-dotenv==0.20.0